NIAD-932: Implemented CR changes
