+++ service/src/main/java/uk/nhs/adaptors/gp2gp/common/configuration/AppInitializer.java
@@ -2,6 +2,7 @@ package uk.nhs.adaptors.gp2gp.common.configuration;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import uk.nhs.adaptors.gp2gp.common.storage.StorageConnectorConfiguration;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.InitializingBean;
@@ -12,15 +13,18 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class AppInitializer implements InitializingBean {
-    private final AppConfiguration appConfiguration;
+    private final StorageConnectorConfiguration storageConnectorConfiguration;
     private final CustomTrustStore customTrustStore;
 
     @Override
     public void afterPropertiesSet() {
         LOGGER.info("Running app initializer");
-        if (StringUtils.isNotBlank(appConfiguration.getTrustStoreUrl())) {
+        if (StringUtils.isNotBlank(storageConnectorConfiguration.getTrustStoreUrl())) {
             LOGGER.info("Adding custom TrustStore to default one");
-            customTrustStore.addToDefault(appConfiguration.getTrustStoreUrl(), appConfiguration.getTrustStorePassword());
+            customTrustStore.addToDefault(storageConnectorConfiguration.getTrustStoreUrl(),
+                storageConnectorConfiguration.getTrustStorePassword());
+        } else {
+            LOGGER.warn("Trust store URL is not set. Running service without the trust store.");
         }
     }
 }
